Refactor: Improve file parsing logic
