+++ src/parser/mod.rs
@@ -1,10 +1,8 @@
 use nom::branch::alt;
-use nom::bytes::complete::take_while1;
-use nom::bytes::tag;
 use nom::character::complete::multispace0;
 use nom::combinator::{complete, map};
-use nom::multi::{many0, separated_list0};
-use nom::sequence::{delimited, preceded};
+use nom::multi::many0;
+use nom::sequence::preceded;
 use nom::{IResult, Parser};
 use log::{error,info};
 
@@ -25,29 +23,6 @@ pub fn is_ident_char(c:char) -> bool{
     c.is_alphanumeric() || c == '_' || c >= '\u{4E00}'
 }
 
-pub fn parse_ident(input:&str) -> IResult<&str,&str> {
-    take_while1(is_ident_char)(input)
-}
-
-pub fn parse_type(input:&str) -> IResult<&str,&str> {
-    take_while1(|c:char| c.is_alphanumeric())(input)
-}
-
-pub fn parse_attr(input:&str) -> IResult<&str,Vec<String>>{
-    let sep = preceded(
-                whitespace0,
-                take_while1(|c:char| c != ',' && c != ']')
-                );
-    let map_fn = map(sep, |s:&str| s.trim().to_string());
-
-    let sep_list = separated_list0(tag(","), map_fn); 
-    let mut parser = delimited(
-        tag("["), 
-        sep_list,
-        tag("]")
-    );
-    parser.parse(input)
-}
 
 pub fn parse_definition(input:&str) -> IResult<&str,GlobalDefinition>{
     let mut parser = preceded(
